+++ example/lib/example5_themes_playground/theme/flex_theme_dark.dart
@@ -154,6 +154,8 @@ FlexColorScheme flexColorSchemeDark(ThemeController controller, Color source) {
             segmentedButtonSchemeColor: controller.segmentedButtonSchemeColor,
             segmentedButtonUnselectedSchemeColor:
                 controller.segmentedButtonUnselectedSchemeColor,
+            segmentedButtonUnselectedForegroundSchemeColor:
+                controller.segmentedButtonUnselectedForegroundSchemeColor,
             segmentedButtonBorderSchemeColor:
                 controller.segmentedButtonBorderSchemeColor,
             segmentedButtonBorderWidth: controller.segmentedButtonBorderWidth,