+++ example/lib/example5_themes_playground/theme/flex_theme_light.dart
@@ -199,6 +199,8 @@ FlexColorScheme flexColorSchemeLight(ThemeController controller, Color source) {
             segmentedButtonSchemeColor: controller.segmentedButtonSchemeColor,
             segmentedButtonUnselectedSchemeColor:
                 controller.segmentedButtonUnselectedSchemeColor,
+            segmentedButtonUnselectedForegroundSchemeColor:
+                controller.segmentedButtonUnselectedForegroundSchemeColor,
             segmentedButtonBorderSchemeColor:
                 controller.segmentedButtonBorderSchemeColor,
             segmentedButtonBorderWidth: controller.segmentedButtonBorderWidth,