+++ example/lib/example5_themes_playground/widgets/panels/menu_settings/menu_settings.dart
@@ -209,15 +209,15 @@ class MenuSettings extends StatelessWidget {
           subtitle: Text('The PopupMenuButton can be used on any kind of '
               'widget, below a custom TextButton is being used.'),
         ),
-        const Padding(
-          padding: EdgeInsets.symmetric(horizontal: 16),
+        Padding(
+          padding: const EdgeInsets.symmetric(horizontal: 16),
           // The button used in the PopupMenuShowcase to open the PopupMenu
           // is not a native widget, only the menu is, and it is the one that
           // is themed.
           child: Wrap(
             spacing: 8,
             runSpacing: 8,
-            children: <Widget>[
+            children: const <Widget>[
               PopupMenuButtonShowcase(),
               PopupMenuButtonTilesShowcase(),
             ],