+++ example/lib/example5_themes_playground/widgets/panels/surface_blend_settings/surface_mode_widgets.dart
@@ -241,11 +241,11 @@ List<Widget> _getModeWidget(ColorScheme scheme, [bool allModes = true]) =>
         message: 'High surface\nlow scaffold',
         child: Icon(Icons.layers),
       ),
-      const Tooltip(
+      Tooltip(
         message: 'High scaffold\nlow surface',
         child: Stack(
           alignment: Alignment.center,
-          children: <Widget>[
+          children: const <Widget>[
             Padding(
               padding: EdgeInsets.only(top: 10),
               child: Icon(Icons.layers_outlined),